Fixed edit account issues

+++ lib/models/CRUDModel.dart
@@ -198,10 +198,11 @@ class FirestoreCRUDModel {
   }
 
   Future<void> deleteProvider({ProviderUser provider}) async {
-    for(Chat chat in provider.chats) {
-      await updateChat(chat: Chat(chatid: chat.chatid, isDeleted : true, lastMessage: 'Provider No Longer Exists'));
+    if(provider.chats.isNotEmpty) for(String chatID in provider.chats) {
+      print('Deleting chats');
+      await updateChat(chat: Chat(chatid: chatID, isDeleted : true, lastMessage: 'Provider No Longer Exists'));
     }
-    await FirebaseStorage.instance.ref().child(provider.displayPictureFileName).delete();
+    if(provider.displayPictureFileName!=null) await FirebaseStorage.instance.ref().child(provider.displayPictureFileName).delete();
     await _providersCollectionRef.document(provider.pid).delete();
     await _usersCollectionRef.document(provider.uid).setData({'providerProfiles': FieldValue.arrayRemove([provider.pid])},merge: true);
   }
